requirements.txt
# Microsoft Agent Framework (PREVIEW - requires --pre flag)
agent-framework-azure-ai --pre

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Azure Services
azure-identity==1.19.0
azure-ai-inference==1.0.0b6
azure-search-documents==11.6.0b7
azure-core==1.31.0

# Vector Database
pinecone-client==5.0.1
langchain-pinecone==0.2.0

# LangChain Components
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langchain-core==0.3.15

# Document Processing
pypdf==5.0.1
pypdf2==3.0.1

# Environment & Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# CORS
fastapi-cors==0.0.6

# Logging
colorlog==6.9.0

# Type hints
typing-extensions==4.12.2
